change names of test files for Maven test
