+++ src/test/java/krypto/lea/LeaTestWithCBCTest.java
@@ -1,16 +1,17 @@
-package krypto.seed;
+package krypto.lea;
 
 import krypto.Algorithm;
-import krypto.SeedAlgorithmFactory;
+import krypto.LeaAlgorithmFactory;
+import krypto.exception.WrongFixedKeySize;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.security.InvalidKeyException;
 
-public class SeedTestWithECB {
+public class LeaTestWithCBCTest {
     @Test
-    void plainTextLessThan16() throws InvalidKeyException {
-        Algorithm algo = SeedAlgorithmFactory.createWithECB("masterkey");
+    void plainTextLessThan16() throws InvalidKeyException, WrongFixedKeySize {
+        Algorithm algo = LeaAlgorithmFactory.createWithCBC("masterkey", 256, "1234567812345678".getBytes());
         String plain = "Hello 안녕";
         byte[] cipherText = algo.encrypt(plain.getBytes());
         byte[] decryptedText = algo.decrypt(cipherText);
@@ -18,8 +19,8 @@ public class SeedTestWithECB {
     }
 
     @Test
-    void shortPlainText() throws InvalidKeyException {
-        Algorithm algo = SeedAlgorithmFactory.createWithECB("masterkey");
+    void shortPlainText() throws InvalidKeyException, WrongFixedKeySize {
+        Algorithm algo = LeaAlgorithmFactory.createWithCBC("masterkey", 256, "1234567812345678".getBytes());
         String plain = "Hello";
         byte[] cipherText = algo.encrypt(plain.getBytes());
         byte[] decryptedText = algo.decrypt(cipherText);
@@ -27,8 +28,8 @@ public class SeedTestWithECB {
     }
 
     @Test
-    void longPlainText() throws InvalidKeyException {
-        Algorithm algo = SeedAlgorithmFactory.createWithECB("masterkey");
+    void longPlainText() throws InvalidKeyException, WrongFixedKeySize {
+        Algorithm algo = LeaAlgorithmFactory.createWithCBC("masterkey", 256, "1234567812345678".getBytes());
         String plain = "Hello 안녕 Bonjour Hola こんにちは 你好";
         StringBuilder repeatedText = new StringBuilder();
         for (int i = 0; i < 100; i++) {
