+++ src/test/java/krypto/lea/LeaTestWithECBTest.java
@@ -1,17 +1,17 @@
-package krypto.seed;
+package krypto.lea;
 
 import krypto.Algorithm;
-import krypto.SeedAlgorithmFactory;
-import krypto.exception.WrongInitialVectorException;
+import krypto.LeaAlgorithmFactory;
+import krypto.exception.WrongFixedKeySize;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.security.InvalidKeyException;
 
-public class SeedTestWithCBC {
+public class LeaTestWithECBTest {
     @Test
-    void plainTextLessThan16() throws InvalidKeyException, WrongInitialVectorException {
-        Algorithm algo = SeedAlgorithmFactory.createWithCBC("masterkey", "1234567812345678".getBytes());
+    void plainTextLessThan16() throws InvalidKeyException, WrongFixedKeySize {
+        Algorithm algo = LeaAlgorithmFactory.createWithECB("masterkey", 256);
         String plain = "Hello 안녕";
         byte[] cipherText = algo.encrypt(plain.getBytes());
         byte[] decryptedText = algo.decrypt(cipherText);
@@ -19,8 +19,8 @@ public class SeedTestWithCBC {
     }
 
     @Test
-    void shortPlainText() throws InvalidKeyException, WrongInitialVectorException {
-        Algorithm algo = SeedAlgorithmFactory.createWithCBC("masterkey", "1234567812345678".getBytes());
+    void shortPlainText() throws InvalidKeyException, WrongFixedKeySize {
+        Algorithm algo = LeaAlgorithmFactory.createWithECB("masterkey", 256);
         String plain = "Hello";
         byte[] cipherText = algo.encrypt(plain.getBytes());
         byte[] decryptedText = algo.decrypt(cipherText);
@@ -28,8 +28,8 @@ public class SeedTestWithCBC {
     }
 
     @Test
-    void longPlainText() throws InvalidKeyException, WrongInitialVectorException {
-        Algorithm algo = SeedAlgorithmFactory.createWithCBC("masterkey", "1234567812345678".getBytes());
+    void longPlainText() throws InvalidKeyException, WrongFixedKeySize {
+        Algorithm algo = LeaAlgorithmFactory.createWithECB("masterkey", 256);
         String plain = "Hello 안녕 Bonjour Hola こんにちは 你好";
         StringBuilder repeatedText = new StringBuilder();
         for (int i = 0; i < 100; i++) {
