+++ src/test/java/krypto/seed/SeedTestWithCBCTest.java
@@ -1,17 +1,17 @@
-package krypto.lea;
+package krypto.seed;
 
 import krypto.Algorithm;
-import krypto.LeaAlgorithmFactory;
-import krypto.exception.WrongFixedKeySize;
+import krypto.SeedAlgorithmFactory;
+import krypto.exception.WrongInitialVectorException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.security.InvalidKeyException;
 
-public class LeaTestWithCBC {
+public class SeedTestWithCBCTest {
     @Test
-    void plainTextLessThan16() throws InvalidKeyException, WrongFixedKeySize {
-        Algorithm algo = LeaAlgorithmFactory.createWithCBC("masterkey", 256, "1234567812345678".getBytes());
+    void plainTextLessThan16() throws InvalidKeyException, WrongInitialVectorException {
+        Algorithm algo = SeedAlgorithmFactory.createWithCBC("masterkey", "1234567812345678".getBytes());
         String plain = "Hello 안녕";
         byte[] cipherText = algo.encrypt(plain.getBytes());
         byte[] decryptedText = algo.decrypt(cipherText);
@@ -19,8 +19,8 @@ public class LeaTestWithCBC {
     }
 
     @Test
-    void shortPlainText() throws InvalidKeyException, WrongFixedKeySize {
-        Algorithm algo = LeaAlgorithmFactory.createWithCBC("masterkey", 256, "1234567812345678".getBytes());
+    void shortPlainText() throws InvalidKeyException, WrongInitialVectorException {
+        Algorithm algo = SeedAlgorithmFactory.createWithCBC("masterkey", "1234567812345678".getBytes());
         String plain = "Hello";
         byte[] cipherText = algo.encrypt(plain.getBytes());
         byte[] decryptedText = algo.decrypt(cipherText);
@@ -28,8 +28,8 @@ public class LeaTestWithCBC {
     }
 
     @Test
-    void longPlainText() throws InvalidKeyException, WrongFixedKeySize {
-        Algorithm algo = LeaAlgorithmFactory.createWithCBC("masterkey", 256, "1234567812345678".getBytes());
+    void longPlainText() throws InvalidKeyException, WrongInitialVectorException {
+        Algorithm algo = SeedAlgorithmFactory.createWithCBC("masterkey", "1234567812345678".getBytes());
         String plain = "Hello 안녕 Bonjour Hola こんにちは 你好";
         StringBuilder repeatedText = new StringBuilder();
         for (int i = 0; i < 100; i++) {
